solution/requirements.txt
setuptools==69.5.1
Pillow
deepspeed
einops
safetensors
timm
accelerate>=0.27.2
natsort>=8.4.0
pydantic==1.10.17
https://github.com/Dao-AILab/flash-attention/releases/download/v2.6.1/flash_attn-2.6.1+cu118torch2.2cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
git+https://github.com/modelscope/data-juicer@2dfd318082a4f257ecf86bbaf387ffffbcc05532
decord
datasets
numpy==1.23.5
opencv-python
omegaconf
imageio[ffmpeg]
imageio[pyav]
tensorboard
diffusers==0.27.0
transformers==4.40.1
func_timeout
sentencepiece
scenedetect[opencv]
sk-video
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
xformers==0.0.24